pyproject.toml
[project]
name = "finance-portfolio"
version = "0.1.0"
description = ""
authors = [
    {name = "Egor Pankov",email = "pankovegor9@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.124.2,<0.125.0)",
    "uvicorn[standart] (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pg8000 (>=1.31.5,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
